Move secrets-config proxy code to options package
